Implement Zobrist hashing and fix critical search issues
- Add Zobrist hash tables for position tracking (data.v)
- Implement incremental hash updates in make_move/takeback (move.v)
- Add proper repetition detection using zobrist hashes (search.v)
- Fix PV following logic to match TSCP behavior (search.v)
- Add gen_caps() for quiescence search optimization (movegen.v)
- Add depth safety checks to prevent stack overflow
- Fix hash field type from int to u64 (defs.v)
- Update hash_key() to return board hash (uci.v)

Improvements: ~192 Elo gain from -288 to -96 vs TSCP

+++ src/defs.v
@@ -60,7 +60,7 @@ pub mut:
 	castle  int
 	ep      int
 	fifty   int
-	hash    int
+	hash    u64
 }
 
 pub fn (m Move) str() string {
